backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15

psycopg2-binary>=2.9

dj-database-url>=2.1
python-dotenv>=1.0

gunicorn>=21.2
whitenoise>=6.6